Address Thomas' comments
