+++ compiler/src/iree/compiler/Codegen/Common/TransformExtensions/CommonExtensions.h
@@ -32,11 +32,6 @@ class TransformTypeInterface;
 
 namespace iree_compiler {
 namespace IREE {
-
-namespace LinalgExt {
-class AttentionOp;
-}  // namespace LinalgExt
-
 namespace transform_dialect {
 /// Selected patterns for ApplyPatternOp.
 struct ApplyPatternsOpPatterns {
