+++ llvm-external-projects/iree-dialects/BUILD.bazel
@@ -377,6 +377,7 @@ cc_library(
     deps = [
         ":IREEDialectsTransforms",
         ":IREELinalgExtDialect",
+        ":IREELinalgExtPasses",
         ":IREELinalgExtTransformOpsIncGen",
         ":IREELinalgExtTransforms",
         ":IREELinalgTransformDialect",
