+++ llvm-external-projects/iree-dialects/lib/Dialect/LinalgExt/TransformOps/CMakeLists.txt
@@ -8,8 +8,9 @@ add_mlir_library(IREELinalgExtTransformOps
   IREEDialectsTransforms
   MLIRRewrite
   MLIRTransformDialect
-  
+
   IREELinalgExtDialect
+  IREELinalgExtPasses
   IREELinalgExtTransforms
 
   MLIRPDLDialect
